move try method from utils to categorical
